10월 31일 s3 파일 테스트, star 수정

+++ src/main/java/org/example/webtoonepics/webtoon/dto/ReviewRequest.java
@@ -15,7 +15,7 @@ public class ReviewRequest {
     private Long webtoonId;
     private Long userInfo;
     private String content;
-    private short star;
+    private Double star;
     private String provider_id;
 
     public Review toEntity(Webtoon webtoon, User user) {
